+++ packages/cli/package.json
@@ -5,11 +5,11 @@
     "code-pushup": "src/bin.js"
   },
   "dependencies": {
-    "chalk": "^5.3.0",
-    "yargs": "^17.7.2",
-    "zod": "^3.22.1",
     "@code-pushup/models": "*",
     "@code-pushup/core": "*",
-    "@code-pushup/utils": "*"
+    "@code-pushup/utils": "*",
+    "yargs": "^17.7.2",
+    "zod": "^3.22.1",
+    "chalk": "^5.3.0"
   }
 }
