+++ packages/cli/src/bin.ts
@@ -1,5 +1,5 @@
 #! /usr/bin/env node
-import { cli } from './index';
+import { cli } from './lib/cli';
 import { hideBin } from 'yargs/helpers';
 
 // bootstrap yargs; format arguments
