+++ packages/core/package.json
@@ -2,10 +2,10 @@
   "name": "@code-pushup/core",
   "version": "0.0.1",
   "dependencies": {
-    "@quality-metrics/models": "*",
-    "@quality-metrics/utils": "*",
-    "chalk": "^5.3.0",
-    "@code-pushup/portal-client": "^0.1.2"
+    "@code-pushup/models": "*",
+    "@code-pushup/utils": "*",
+    "@code-pushup/portal-client": "^0.1.2",
+    "chalk": "^5.3.0"
   },
   "type": "commonjs",
   "main": "./index.cjs"
