+++ packages/core/src/lib/commands/collect.spec.ts
@@ -1,5 +1,5 @@
-import { reportSchema } from '@quality-metrics/models';
-import { mockCoreConfig } from '@quality-metrics/models/testing';
+import { reportSchema } from '@code-pushup/models';
+import { mockCoreConfig } from '@code-pushup/models/testing';
 import { describe, expect, it } from 'vitest';
 import { CollectOptions, collect } from './collect';
