+++ packages/core/src/lib/implementation/json-to-gql.ts
@@ -3,7 +3,7 @@ import {
   IssueSourceType,
   SaveReportMutationVariables,
 } from '@code-pushup/portal-client';
-import { Issue, Report } from '@quality-metrics/models';
+import { Issue, Report } from '@code-pushup/models';
 
 export function jsonToGql(report: Report) {
   return {
