+++ packages/core/src/lib/implementation/persist.spec.ts
@@ -1,13 +1,13 @@
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { logPersistedResults, persistReport } from './persist';
 import { readFileSync, unlinkSync } from 'fs';
-import { Report } from '@quality-metrics/models';
+import { Report } from '@code-pushup/models';
 import {
   dummyConfig,
   dummyReport,
   MEMFS_VOLUME,
   mockPersistConfig,
-} from '@quality-metrics/models/testing';
+} from '@code-pushup/models/testing';
 import { vol } from 'memfs';
 import { join } from 'path';
 import { mockConsole, unmockConsole } from '../../../test/console.mock';
