+++ packages/core/test/base.command.mock.ts
@@ -1,4 +1,4 @@
-import { mockCoreConfig } from '@quality-metrics/models/testing';
+import { mockCoreConfig } from '@code-pushup/models/testing';
 import { CommandBaseOptions } from '../src/lib/implementation/model';
 
 export function commandBaseOptionsMock(): CommandBaseOptions {
