+++ packages/models/package.json
@@ -2,7 +2,7 @@
   "name": "@code-pushup/models",
   "version": "0.0.1",
   "dependencies": {
-    "@code-pushup/portal-client": "^0.1.2",
-    "zod": "^3.22.1"
+    "zod": "^3.22.1",
+    "@code-pushup/portal-client": "^0.1.2"
   }
 }
