+++ packages/nx-plugin/package.json
@@ -2,8 +2,8 @@
   "name": "@code-pushup/nx-plugin",
   "version": "0.0.1",
   "dependencies": {
-    "@nx/devkit": "^16.8.1",
-    "tslib": "2.6.2"
+    "tslib": "*",
+    "@nx/devkit": "^16.8.1"
   },
   "type": "commonjs",
   "main": "./src/index.js",
