+++ packages/plugin-eslint/src/lib/eslint-plugin.ts
@@ -1,5 +1,5 @@
-import { AuditOutputs, PluginConfig } from '@quality-metrics/models';
-import { objectToCliArgs } from '@quality-metrics/utils';
+import { AuditOutputs, PluginConfig } from '@code-pushup/models';
+import { objectToCliArgs } from '@code-pushup/utils';
 import * as eslint from 'eslint';
 
 type ESLintPluginConfig = {
