+++ packages/utils/src/lib/report-to-md.spec.ts
@@ -4,7 +4,7 @@ import {
   dummyReport,
   lighthouseReport,
   nxValidatorsOnlyReport,
-} from '@quality-metrics/models/testing';
+} from '@code-pushup/models/testing';
 
 describe('report-to-md', () => {
   it('should contain all sections when using dummy report', () => {
